requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.1.15
django-allauth==0.42.0
gunicorn==20.0.4
idna==2.10
isort==5.5.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
pycodestyle==2.6.0
pylint==2.6.0
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
Unipath==1.1
urllib3==1.25.10
whitenoise==5.2.0
wrapt==1.12.1
djangorestframework-datatables==0.5.2